refactor: move mutateSerializeMdx to /lib
